Rename the audio-s5gen2 plugin to qc-s5gen2
Qualcomm can't possibly have a problem with the two letters of 'qc', and I'd
argue even qualcomm-s5gen2 is 'fair use'.
