+++ contrib/ci/check-source.py
@@ -87,14 +87,6 @@ class Checker:
         valid_prefixes = []
         valid_prefixes.append(prefix)
         for key, value in {
-            "fu_audio_s5gen2_device": "fu_qc_s5gen2_device",
-            "fu_audio_s5gen2_hid_device": "fu_qc_s5gen2_hid_device",
-            "fu_audio_s5gen2_firmware": "fu_qc_s5gen2_firmware",
-            "fu_audio_s5gen2": "fu_qc_s5gen2",  # FIXME: rename after merging #7638
-            "fu_audio_s5gen2_hid": "fu_qc_s5gen2_hid",
-            "fu_audio_s5gen2_ble": "fu_qc_s5gen2_ble",
-            "fu_audio_s5gen2_hid_device": "fu_qc_s5gen2_hid_device",
-            "fu_audio_s5gen2_ble_device": "fu_qc_s5gen2_ble_device",
             "fu_crc": "fu_misr",  # FIXME: split out to fu-misr.[c|h]
             "fu_darwin_efivars": "fu_efivars",
             "fu_dbus_daemon": "fu_daemon",
