+++ plugins/meson.build
@@ -28,7 +28,6 @@ plugins = {
   'analogix': false,
   'android-boot': false,
   'ata': false,
-  'audio-s5gen2': false,
   'aver-hid': false,
   'bcm57xx': false,
   'bios': false,
@@ -96,6 +95,7 @@ plugins = {
   'pci-psp': false,
   'pixart-rf': false,
   'powerd': false,
+  'qc-s5gen2': false,
   'qsi-dock': false,
   'realtek-mst': false,
   'redfish': false,
