+++ plugins/qc-s5gen2/fu-qc-s5gen2.rs
